refactor(#15): 등록일시 데이트 포맷 공통함수 적용
